pyproject.toml
[project]
name = "multi-agent-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.5",
    "prometheus-client>=0.23.1",
    "python-dotenv>=1.2.1",
]
